requirements.txt
django
gunicorn
pylint
black
autopep8
flake8
tox
line-bot-sdk
requests
beautifulsoup4